refactor: Account 도메인 sign-up dto 리팩토링
Account 도메인의 sign-up controller, service dto를 리팩토링한다.

Resolves: #20
