+++ src/modules/account/api/dto/req/sign-up.req.ts
@@ -1,0 +1,38 @@
+import { IsEmail, IsNotEmpty, Matches, IsString } from 'class-validator';
+import { MatchProperty } from '../../../../../common/decorators/match-property.decorator';
+import { SignUpServiceDto } from './../../../application/dto/sign-up-service.dto';
+
+export class SignUpReq {
+  @IsEmail(
+    {},
+    { message: '이메일 형식이 아닙니다.\n이메일을 다시 입력하세요.' },
+  )
+  @IsNotEmpty({ message: '이메일을 입력하세요.' })
+  public readonly email: string;
+
+  @Matches(/^.*(?=.{8,20})(?=.*[0-9])(?=.*[a-zA-Z]).*$/, {
+    message: '비밀번호는 영문+숫자 포함 8~20자리입니다.',
+  })
+  @IsNotEmpty({ message: '비밀번호를 입력하세요.' })
+  public readonly password: string;
+
+  @MatchProperty('password')
+  @Matches(/^.*(?=.{8,20})(?=.*[0-9])(?=.*[a-zA-Z]).*$/, {
+    message: '비밀번호는 영문+숫자 포함 8~20자리입니다.',
+  })
+  @IsNotEmpty({ message: '비밀번호를 입력하세요.' })
+  public readonly passwordCheck: string;
+
+  @IsString({ message: '이름을 입력하세요.' })
+  @IsNotEmpty({ message: '이름을 입력하세요.' })
+  public readonly name: string;
+
+  public toServiceDto(): SignUpServiceDto {
+    const dto = new SignUpServiceDto();
+    dto.email = this.email;
+    dto.name = this.name;
+    dto.password = this.password;
+
+    return dto;
+  }
+}
