+++ src/modules/account/application/dto/index.ts
@@ -1,0 +1,3 @@
+import { SignUpServiceDto } from './sign-up-service.dto';
+
+export { SignUpServiceDto };
